rename : package명 세분화 rename : Util에 있던 DummyData -> FakeData.kt 추가후 이동 remove : MutiMode 삭제 후 Multi-Type들 Adapter로 이동및 companion object로 수행
