+++ app/build.gradle.kts
@@ -41,7 +41,7 @@ android {
 
 dependencies {
 
-    implementation("androidx.activity:activity-ktx:1.9.0")
+//    implementation("androidx.activity:activity-ktx:1.9.0")
 
     implementation("androidx.core:core-ktx:1.12.0")
     implementation("androidx.appcompat:appcompat:1.6.1")
