+++ app/src/main/java/com/example/nbc_standard_multi_view/ui/MainActivity.kt
@@ -1,12 +1,13 @@
-package com.example.nbc_standard_multi_view
+package com.example.nbc_standard_multi_view.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.nbc_standard_multi_view.util.dummyData
-import com.example.nbc_standard_multi_view.util.dummyItems
-import com.example.nbc_standard_multi_view.util.mainMoneyFormat
+import com.example.nbc_standard_multi_view.data.dummyData
+import com.example.nbc_standard_multi_view.data.dummyItems
+import com.example.nbc_standard_multi_view.data.mainMoneyFormat
 import com.example.nbc_standard_multi_view.databinding.ActivityMainBinding
+import com.example.nbc_standard_multi_view.ui.CardAdapter
 
 class MainActivity : AppCompatActivity() {
 
@@ -14,15 +15,21 @@ class MainActivity : AppCompatActivity() {
         ActivityMainBinding.inflate(layoutInflater)
     }
 
+    private lateinit var cardAdapter : CardAdapter
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
 
         dummyData()
-
-        binding.rvMain.adapter = CardAdapter(dummyItems)
-        binding.rvMain.layoutManager = LinearLayoutManager(this)
+        setUpRecyclerView()
 
         binding.tvMainMoney.text = mainMoneyFormat(285856.20F)
+    }
+
+    private fun setUpRecyclerView() {
+        cardAdapter = CardAdapter(dummyItems)
+        binding.rvMain.adapter = cardAdapter
+        binding.rvMain.layoutManager = LinearLayoutManager(this)
     }
 }
