--- app/src/main/java/com/example/nbc_standard_multi_view/util/MutiMode.kt
@@ -1,5 +1,0 @@
-package com.example.nbc_standard_multi_view.util
-
-const val MULTI_TYPE_1 = 1
-const val MULTI_TYPE_2 = 2
-const val MULTI_TYPE_3 = 3
